+++ user-frontend/assets/css/main.css
@@ -44,3 +44,14 @@
   line-height: 20px;
   color: #d5d5d5;
 }
+
+.loading-skeleton {
+  width: 100%;
+  height: 500px;
+  display: flex;
+  flex-direction: column;
+  -webkit-box-pack: center;
+  justify-content: center;
+  -webkit-box-align: center;
+  align-items: center;
+}
