+++ user-frontend/components/Index/BaseDashboard.vue
