+++ user-frontend/components/Session/BaseQuestionDetails.vue
