+++ user-frontend/pages/session/_id.vue
@@ -8,51 +8,8 @@
     <v-row class="fill-height">
       <v-col md="12" lg="3">
         <div class="sticky-column mx-5 ml-lg-5">
-          <BaseQuestion :question="match_question" />
-          <v-card outlined class="mt-3 white--text" color="#242424">
-            <v-card-title> Match Details</v-card-title>
-            <v-card-text class="white--text">
-              <v-icon color="white" class="mr-1"
-                >mdi-shield-sword-outline</v-icon
-              >
-              Match Mode:
-              <v-chip
-                small
-                :color="mode_chip_colors[match.mode]"
-                text-color="white"
-              >
-                {{ match.mode }}
-              </v-chip>
-
-              <br />
-
-              <v-countdown
-                v-if="match.match_requirements.question_mode === 'timed'"
-                :left-time="time_left"
-                @finish="endMatch"
-              >
-                <template #process="{ timeObj }">
-                  <v-icon color="white" class="mr-1">mdi-timer</v-icon>
-                  Time Left: <b>{{ `${timeObj.m}:${timeObj.s}` }}</b>
-                  <br />
-                </template>
-              </v-countdown>
-              <v-divider class="my-6 white"></v-divider>
-              <v-icon color="white" class="mr-1">mdi-account</v-icon>
-              Partner's Display Name: <b>{{ partner_info.display_name }}</b>
-              <br />
-              <v-icon color="white" class="mr-1">mdi-code-tags</v-icon>
-              Programming Language:
-              <b class="text-capitalize">{{
-                match.match_requirements.programming_language
-              }}</b>
-            </v-card-text>
-            <v-card-actions class="pb-6 mx-2">
-              <v-btn color="error" block depressed @click="endMatch">
-                End Match
-              </v-btn>
-            </v-card-actions>
-          </v-card>
+          <BaseQuestionDetails :question="match_question" />
+          <BaseMatchDetails @end-match="endMatch" />
         </div>
       </v-col>
       <v-col md="12" lg="6">
@@ -70,15 +27,18 @@
 import systemMixin from "@/mixins/system";
 import matchMixin from "@/mixins/match";
 
-import BaseChat from "@/components/Match/BaseChat";
-import BaseCodeEditor from "@/components/Match/BaseCodeEditor";
-import BaseQuestion from "@/components/Match/BaseQuestion";
+import BaseQuestionDetails from "@/components/Session/BaseQuestionDetails";
+import BaseMatchDetails from "@/components/Session/BaseMatchDetails";
+import BaseChat from "@/components/Session/BaseChat";
+import BaseCodeEditor from "@/components/Session/BaseCodeEditor";
 
 export default {
+  name: "Match",
   components: {
+    BaseQuestionDetails,
+    BaseMatchDetails,
     BaseChat,
     BaseCodeEditor,
-    BaseQuestion,
   },
   mixins: [matchMixin, systemMixin],
   beforeRouteLeave(to, from, next) {
@@ -101,19 +61,20 @@ export default {
     return {
       match_id: undefined,
       match_question: undefined,
-      show: false,
       match_ended: false,
     };
   },
   async fetch() {
     try {
       this.SET_LOADING({ data: true });
-      this.match_id = localStorage.getItem("match_id") || this.$route.params.id; // Either from localStorage or URL params
+      // Retrieve Match ID either from localStorage or URL params
+      this.match_id = localStorage.getItem("match_id") || this.$route.params.id;
       // If match is not in store, retrieve from server
       if (!this.match) {
         await this.GET_MATCH({ match_id: this.match_id });
       }
-      this.match_question = _.get(this.match, "question"); // Set it to data to be passed to question component
+      // Set it to data to be passed to question component
+      this.match_question = _.get(this.match, "question");
       // this.checkIsValidMatch();
     } catch (err) {
       console.error(err);
@@ -122,32 +83,10 @@ export default {
       this.SET_LOADING({ data: false });
     }
   },
-  computed: {
-    partner_info() {
-      const partner1_id = _.get(this.match, "partner1._id");
-      if (partner1_id === this.user_id) {
-        return this.match.partner1;
-      }
-      return this.match.partner2;
-    },
-    /**
-     * @description Computes the exact duration left
-     */
-    time_left() {
-      const updated_at = _.get(this.match, "updated_at");
-      const recommended_duration = _.get(
-        this.match_question,
-        "recommended_duration",
-      );
-      const end_time = this.$moment(updated_at).add(
-        recommended_duration,
-        "minutes",
-      );
-      const time_left = end_time.diff(this.$moment(), "milliseconds");
-      return time_left;
-    },
-  },
   methods: {
+    /**
+     * @description Check if the match is valid based on the ID
+     */
     async checkIsValidMatch() {
       const is_match_ended = _.get(this.match, "status") === "completed";
       if (is_match_ended) {
@@ -211,6 +150,9 @@ export default {
         return;
       }
     },
+    /**
+     * @description Ends the page and re-route user to thank you page
+     */
     async endMatch() {
       try {
         this.SET_LOADING({ data: true });
